ajuste de que borraba tambn la ausencia si habia en la semana cuando se modificaba los cuadrantes

+++ src/cuadrantes/cuadrantes.mongodb.ts
@@ -120,10 +120,24 @@ export class CuadrantesDatabase {
     const db = (await this.mongoDbService.getConexion()).db("soluciones");
     const cuadrantesCollection = db.collection<TCuadrante>("cuadrantes2");
 
+    // Verificar si el turno tiene una ausencia asociada
+    const turnoConAusencia = await cuadrantesCollection.findOne({
+      idPlan: idPlan,
+      "ausencia.tipo": { $exists: true }, // Verifica si existe el campo ausencia.tipo
+    });
+
+    if (turnoConAusencia) {
+      console.log(
+        `No se puede borrar el turno ya que tiene una ausencia asociada con el idPlan: ${idPlan}`,
+      );
+      return false;
+    }
+
+    // Si no tiene ausencia, proceder a borrar
     const resCuadrantes = await cuadrantesCollection.deleteOne({
       idPlan: idPlan,
     });
-
+    console.log(resCuadrantes);
     return resCuadrantes.acknowledged;
   }
